Dont hard-code `:current_user` as resource key, update docs to reference change
Originally this was planned to work with a current user
to be expected, but with the authentication approach of
the various libaries many use in the ecosystem dont
necessarily default to this often. Therefore, we have
it default over to `:current_user` if it is not set in
the environment.